move encoders to Internal or TimeSeriesApp objects
it cleans up the code from unneccessary noise
